refactor: move OutgoingMessageRecipientsRetriever.kt to `domain` folder
